Added logic for manually parsing the text

+++ src/main/java/com/mk/parser/model/ParsedData.java
@@ -4,16 +4,18 @@ import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @ToString
 public class ParsedData {
     private String extractedText;
-    private long extractedValue;
+    private BigDecimal extractedValue;
     private int startPosition;
     private int endPosition;
 
-    public ParsedData(String extractedText, long extractedValue, int startPosition, int endPosition) {
+    public ParsedData(String extractedText, BigDecimal extractedValue, int startPosition, int endPosition) {
         this.extractedText = extractedText;
         this.extractedValue = extractedValue;
         this.startPosition = startPosition;
